requirements.txt
streamlit
google-api-python-client
python-dotenv
requests
pandas
openai
numpy
altair
matplotlib


